feat: UI overhaul, model catalog, emotes, plugin config system
- Dynamic model catalog: fetch models from OpenAI, Anthropic, Groq,
  xAI, OpenRouter, Google, Ollama, Vercel AI Gateway with per-provider
  caching and searchable dropdown selection for all MODEL fields
- Plugin UI config rendering system with typed field catalog
- Page reorganization: Settings, Agent, Database, Logs, Config,
  Media Gallery, Vector Browser, and Plugins page views
- Emote system for agent avatar animations
- Autonomy source channel labels
- Enhanced SkillsView with source editing
- CharacterView overhaul
- Custom searchable dropdown replacing HTML5 datalist
- Fetch Models button in Settings for on-demand refresh
- Partial config save validation fix
- Inline plugin config blocks in chat

+++ src/api/database.ts
@@ -399,7 +399,7 @@ async function handleGetStatus(
 
   const tableResult = await executeRawSql(
     runtime,
-    `SELECT count(*)::int AS cnt
+    `SELECT count(*) AS cnt
        FROM information_schema.tables
       WHERE table_schema NOT IN ('pg_catalog', 'information_schema')
         AND table_type = 'BASE TABLE'`,
@@ -626,7 +626,7 @@ async function handleGetTables(
     `SELECT
        t.table_schema AS schema,
        t.table_name AS name,
-       COALESCE(s.n_live_tup, 0)::int AS row_count
+       COALESCE(s.n_live_tup, 0) AS row_count
      FROM information_schema.tables t
      LEFT JOIN pg_stat_user_tables s
        ON s.schemaname = t.table_schema
@@ -775,7 +775,7 @@ async function handleGetRows(
   // Count total (with search filter)
   const countResult = await executeRawSql(
     runtime,
-    `SELECT count(*)::int AS total FROM ${quoteIdent(tableName)} ${whereClause}`,
+    `SELECT count(*) AS total FROM ${quoteIdent(tableName)} ${whereClause}`,
   );
   const total = Number(
     (countResult.rows[0] as Record<string, unknown>)?.total ?? 0,
